requirements.txt
# ============================================================
# Portafolio Ciencia de Datos — Colombia
# requirements.txt
# Python 3.10+
# ============================================================

# --- Manipulación de datos ---
pandas==2.2.2
numpy==1.26.4
pyarrow==16.0.0          # Lectura/escritura Parquet
duckdb==0.10.3            # Consultas SQL sobre Parquet sin base de datos

# --- Descarga y peticiones HTTP ---
requests==2.31.0
httpx==0.27.0
tqdm==4.66.4              # Barras de progreso en descargas

# --- Archivos SPSS/SAV (formato DANE) ---
pyreadstat==1.2.7         # Leer archivos .sav del DANE

# --- Análisis estadístico ---
scipy==1.13.1
statsmodels==0.14.2       # Regresiones, modelos econométricos
scikit-learn==1.5.0       # Machine learning y modelos predictivos

# --- Visualización ---
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0            # Gráficas interactivas para reportes web
kaleido==0.2.1            # Exportar gráficas Plotly a imagen estática

# --- Jupyter y entorno interactivo ---
jupyterlab==4.2.1
ipywidgets==8.1.3
nbformat==5.10.4

# --- Reportes y documentación ---
jinja2==3.1.4             # Templates para reportes automáticos
tabulate==0.9.0           # Tablas en texto plano y markdown

# --- Utilidades ---
python-dotenv==1.0.1      # Variables de entorno (credenciales, rutas)
loguru==0.7.2             # Logs claros para los pipelines
pyyaml==6.0.1             # Archivos de configuración
click==8.1.7              # Scripts de línea de comandos con argumentos
